test: consolidate grid LitRenderer column ITs into one view
Move the LitRenderer-columns showcase (index variable, computed and
complex withProperty, @click handlers mutating via refreshItem/refreshAll)
from the demo GridViewUsingTemplates page/IT into the dedicated
GridLitRendererPage/IT, then delete the demo pair. These exercise
LitRenderer features not covered by the existing render/switch/editor
tests, so nothing is dropped. Give the existing grid an explicit id so
the reattach test no longer depends on DOM order now that the page hosts
a second grid.
